+++ HISTORY.rst
@@ -4,6 +4,8 @@ Changelog
 1.2.1 - Unreleased
 ------------------
 
+* Allow to specify multiple masters per instance.
+  [fschulze]
 
 
 1.2.0 - 2015-03-05